Create AbstractServlet class and move here common functionality for all Servlets.
