+++ src/main/java/com/example/httplearningapi/controller/webcontroller/servlets/PrescriptionsServlet.java
@@ -1,48 +1,22 @@
 package com.example.httplearningapi.controller.webcontroller.servlets;
 
-import jakarta.servlet.ServletException;
+import com.example.httplearningapi.model.entities.user.Prescription;
+import com.example.httplearningapi.model.service.PrescriptionService;
+import com.example.httplearningapi.model.service.Service;
+import com.example.httplearningapi.util.ExceptionHandleUtil;
 import jakarta.servlet.annotation.WebServlet;
-import jakarta.servlet.http.HttpServlet;
-import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
-import java.io.IOException;
-
 @WebServlet(urlPatterns = {"/prescriptions", "/prescriptions/*"})
-public class PrescriptionsServlet extends HttpServlet {
+public class PrescriptionsServlet extends AbstractServlet {
 
     @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        super.doGet(req, resp);
-    }
-
-    @Override
-    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        super.doPost(req, resp);
-    }
-
-    @Override
-    protected void doPut(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        super.doPut(req, resp);
-    }
-
-    @Override
-    protected void doDelete(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        super.doDelete(req, resp);
-    }
-
-    @Override
-    protected void doHead(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        super.doHead(req, resp);
-    }
-
-    @Override
-    protected void doOptions(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        super.doOptions(req, resp);
-    }
-
-    @Override
-    protected void doTrace(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        super.doTrace(req, resp);
+    void handleRequest(HandleProcessor<Service<?>> handleProcessor, HttpServletResponse resp) {
+        try {
+            Service<Prescription> service = new PrescriptionService();
+            handleProcessor.process(service);
+        } catch (Exception e) {
+            ExceptionHandleUtil.processException(e, resp);
+        }
     }
 }
