+++ src/main/java/com/example/httplearningapi/controller/webcontroller/servlets/UsersServlet.java
@@ -4,53 +4,14 @@ import com.example.httplearningapi.model.entities.user.User;
 import com.example.httplearningapi.model.service.Service;
 import com.example.httplearningapi.model.service.UserService;
 import com.example.httplearningapi.util.ExceptionHandleUtil;
-import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
-import jakarta.servlet.http.HttpServlet;
-import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
-import java.io.IOException;
-
 @WebServlet(urlPatterns = {"/users", "/users/*"})
-public class UsersServlet extends HttpServlet {
+public class UsersServlet extends AbstractServlet {
 
     @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) {
-        this.handleRequest((service) -> service.handleGet(req.getPathInfo(), req, resp), resp);
-    }
-
-    @Override
-    protected void doPost(HttpServletRequest req, HttpServletResponse resp) {
-        this.handleRequest((service) -> service.handlePost(req.getPathInfo(), req, resp), resp);
-    }
-
-    @Override
-    protected void doPut(HttpServletRequest req, HttpServletResponse resp) {
-        this.handleRequest((service) -> service.handlePut(req.getPathInfo(), req, resp), resp);
-    }
-
-    @Override
-    protected void doDelete(HttpServletRequest req, HttpServletResponse resp) {
-        this.handleRequest((service) -> service.handleDelete(req.getPathInfo(), req, resp), resp);
-    }
-
-    @Override
-    protected void doHead(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        super.doHead(req, resp);
-    }
-
-    @Override
-    protected void doOptions(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        super.doOptions(req, resp);
-    }
-
-    @Override
-    protected void doTrace(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        super.doTrace(req, resp);
-    }
-
-    private void handleRequest(HandleProcessor<Service<User>> handleProcessor, HttpServletResponse resp) {
+    void handleRequest(HandleProcessor<Service<?>> handleProcessor, HttpServletResponse resp) {
         try {
             Service<User> service = new UserService();
             handleProcessor.process(service);
